chore: try pass ci

+++ tests/test_ifnet.py
@@ -19,8 +19,8 @@ class Test_IFNet:
             model: VFIBaseModel = AutoModel.from_config(config=cfg, fp16=False, device=get_device())
             print(model.device)
 
-            I0 = torch.from_numpy(img0).permute(2, 0, 1).unsqueeze(0).half().to(model.device) / 255.0
-            I1 = torch.from_numpy(img1).permute(2, 0, 1).unsqueeze(0).half().to(model.device) / 255.0
+            I0 = torch.from_numpy(img0).permute(2, 0, 1).unsqueeze(0).to(model.device) / 255.0
+            I1 = torch.from_numpy(img1).permute(2, 0, 1).unsqueeze(0).to(model.device) / 255.0
             I0 = I0.unsqueeze(0)
             I1 = I1.unsqueeze(0)
             inp = torch.cat([I0, I1], dim=1)
